backend/requirements.txt
# Python dependencies for Claude Code PM backend
# Core web framework for building APIs
fastapi==0.104.1

# ASGI server for running FastAPI
uvicorn[standard]==0.24.0

# HTTP client library for making API calls
requests==2.31.0

# Data validation and settings management
pydantic==2.5.0

# Environment variable management
python-dotenv==1.0.0

# CORS middleware for cross-origin requests
fastapi-cors==0.0.6

# JSON Web Token for authentication
PyJWT==2.8.0

# Cryptography for secure operations
cryptography==42.0.8

# Async HTTP client
httpx==0.25.2

# Logging and monitoring
structlog==23.2.0

# System monitoring
psutil==5.9.6

# Development and testing tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
